% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-apply.R
\name{bin_apply}
\alias{bin_apply}
\alias{bin_apply_all}
\title{Bin a variable (typically time) and compute an aggregate for each bin}
\usage{
bin_apply(data, y, x = t, x_bin_length = mins(30), FUN = mean, ...)

bin_apply_all(data, ...)
}
\arguments{
\item{data}{\link{data.table} or \link{behavr} table (see details)}

\item{y}{variable to be aggregated}

\item{x}{the variable to be binned}

\item{x_bin_length}{the length of the bins (same using as `x``)}

\item{FUN}{function used to aggregate (e.g. \link{mean}, \link{median}, \link{sum} and so on)}

\item{...}{additional arguments to be passed to \code{FUN}}
}
\description{
This function is typically used to summarise (i.e. computing an aggreate of) a variable (\code{y})
for bins of a another variable \code{x} (typically time).
.
}
\details{
\code{bin_apply} expects data from a single individal.
\code{bin_apply_all} works on multiple individuals identifies by a unique key.
}
\examples{
query <- data.frame(experiment_id="toy_experiment",
                      region_id=1:5)
dt <- toy_activity_data(query, duration=days(4))

# average by 30min time bins, default
dt_binned <- bin_apply_all(dt,moving)
# equivalent to
dt_binned <- dt[, bin_apply(.SD, moving),by="id"]

# More advanced usage
dt <- toy_dam_data(query, duration=days(4))

# nsum activity ber 60minutes
dt_binned <- bin_apply_all(dt,
                           activity,
                           x=t,
                           x_bin_length = mins(60),
                           FUN=sum)

}
