% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_behavr}
\alias{merge_behavr}
\title{Merge several behavr tables into a single dataframe}
\usage{
merge_behavr(x, y)
}
\description{
This function merges two behavr tables,
represented by x and y, in such a way
that x gains all the columns in y not present in x
The merge is performed using the t column
}
