% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavr.R
\name{meta}
\alias{meta}
\alias{setmeta}
\title{Retreive metadata}
\usage{
meta(x)

setmeta(x, new)
}
\arguments{
\item{x}{a \link{behavr} object}

\item{new}{a new metadata table}
}
\value{
a \link{data.table} representing the metadata in \code{x}
}
\description{
This function returns the meta data from a \link{behavr} object
}
\examples{
set.seed(1)
met <- data.table::data.table(id = 1:5,
                              condition=letters[1:5],
                              sex=c("M","M","M","F", "F"),
                              key="id")
t <- hms::as.hms(1L:100L)
data <- met[  ,
              list(t=t,
                  x=rnorm(100),
                  y=rnorm(100),
                  eating=runif(100) > .5 ),
              by="id"]

d <- behavr(data,met)
##### Show metadata
meta(d)
class(d)
d2 <- d[id==1]
meta(d2)

##### Alter metadata
# meta(d)[, treatment := interaction(condition,sex)]
}
\seealso{
\link{behavr} to generate a \code{behavr} object, \link{xmd} to map metadata variables to data
}
