% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavr.R
\name{print.behavr}
\alias{print.behavr}
\alias{summary.behavr}
\title{Print and summarise a \link{behavr} table}
\usage{
\method{print}{behavr}(x, ...)

\method{summary}{behavr}(object, ...)
}
\arguments{
\item{x, object}{\link{behavr} table}

\item{...}{arguments passed on to further method}
}
\description{
Print and summarise a \link{behavr} table
}
\seealso{
\itemize{
\item \link{behavr} -- to generate x
\item \link{print.default}
\item \link{summary.default}
}
}
