% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.R
\name{rbind_behavr}
\alias{rbind_behavr}
\title{Rbind two behavrs with same columns in data and metadata}
\usage{
rbind_behavr(x, y)
}
\arguments{
\item{x}{First behavr table}

\item{y}{Second behavr table}
}
\value{
Rbinded behavr
}
\description{
Rbind two behavrs with same columns in data and metadata
}
