% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindlist.R
\name{rbindlist.behavr}
\alias{rbindlist.behavr}
\title{Put together a list of \link{behavr} tables}
\usage{
rbindlist.behavr(l)
}
\arguments{
\item{l}{list of \link{behavr}}
}
\description{
Put together a list of \link{behavr} tables
}
