% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify_behavr}
\alias{simplify_behavr}
\title{Make a writable behavr object}
\usage{
simplify_behavr(data, meta = FALSE)
}
\arguments{
\item{data}{A rejoined behavr or any data.table object}

\item{meta}{If TRUE, repeat also for the metadata}
}
\value{
A behavr table with no columns of type list
TODO Possibly this should be part of a fwrite_behavr function in behavr
}
\description{
Make a writable behavr object
}
