% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\title{Sticth behavioural data by puting together individuals belonging to different experiments}
\usage{
stitch(x, on, time_ref = "date", time_variable = "t")
}
\arguments{
\item{x}{\link{behavr} object}

\item{on}{name of a metavariable serving as a unique id (per individual)}

\item{time_ref}{name of a metavariable used to align time (e.g. \code{"date"}, or \code{"datetime"})}

\item{time_variable}{name of the variable describing time}
}
\value{
a \link{behavr} table
}
\description{
This functions can merge rows of data from the same individual, that was recorded over multiple experiments.
A common scenario is when an experiment is interrupted (the individuals could be, for instcane, shuffled), and a new recording is started.
Stitching assumes the users has defined a \emph{unique id} in the metadata that referes to a spefific individual.
Then, if any data that comes form the same unique id, it is merged.
}
\details{
When several id match a unique id (several experiments), the first (in time) experiment is used to generate the new id.
The data from the following one(s) will be added with a time lag equls to the difference between the \code{time_ref}.
When data is aligned to a circadian time already, it makes sense to use \code{"date"} as time ref. Otherwise, \code{"datetime"} makes more sense.
}
\examples{
set.seed(1)
met1 <- data.table::data.table(uid = 1:5,id = 1:5,
                               condition = letters[1:5],
                               sex=c("M","M","M","F", "F"),
                               key="id")
met2 <- data.table::data.table(uid = 1:4,id = 6:9,
                               condition = letters[1:4],
                               sex=c("M","M","M","F"),
                               key="id")
met1[, date := as.POSIXct("2015-01-02")]
met2[, date := as.POSIXct("2015-01-03")]
met <- rbind(met1, met2)
data.table::setkeyv(met, "id")
t <- 1L:100L
data <- met[,list(t=t, x=rnorm(100),y=rnorm(100), eating=runif(100) > .5 ),by="id"]
d <- behavr(data,met)
summary(d)
d2 <- stitch(d, on ="uid")
summary(d2)

}
\seealso{
\link{behavr} to generate a \code{behavr} object
}
