% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-time.R
\name{time_conversion}
\alias{time_conversion}
\alias{days}
\alias{hours}
\alias{mins}
\title{Time conversion utilities}
\usage{
days(x)

hours(x)

mins(x)
}
\arguments{
\item{x}{Numerical vector to be converted in second}
}
\value{
Number of seconds corresponding to \code{x} (1d = 86400s, 1h = 3600s and 1min = 60s)
}
\description{
Trivial functions to convert time to seconds -- as \code{rethomics} uses second as a conventionnal unit of time.
}
\details{
Given an dummy function that takes time in second like: \code{myFunction(t)},
it is always preferqable to call \code{myFunction(days(1.5))} rather than \code{myFunction(60*60*24*1.5)}.
}
