% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmv.R
\name{xmv}
\alias{xmv}
\title{Extract a metavariable and map it against the data}
\usage{
xmv(var)
}
\arguments{
\item{var}{the variable to be extracted}
}
\value{
a vector of the same type as \code{var}, but of the same length as the number of row in the parent data.
As each row of data is matched against metadata for this specific variable.
}
\description{
This function eXpands a MetaVariable from a parent \link{behavr} object.
That is it matches this variable (from metadata) to the data \emph{by id}.
}
\details{
This function \emph{can only be called within between the \code{[]} of a parent} \link{behavr} object.
It is intended to facilitate operations between data and metadata.
For instance, when one wants to modify a column of the data according a metavariable.
}
\examples{
library(data.table)
set.seed(1)
data <- data.table(
                   id = rep(c("A","B"), times=c(10,26)),
                   t = c(1:10,5:30),
                   x = rnorm(36), key="id"
                   )

metadata = data.table(id=c("A","B"), treatment=c("w","z"), lifespan=c(19,32), ref_x=c(1,0),key="id")
dt <- behavr(data,metadata)
summary(dt)

#### Subsetting using metadata
dt[xmv(treatment) == "w"]
dt[xmv(treatment) == "w"]
dt[xmv(lifespan) < 30]

#### Allocating new columns using metavariable
# Just joining lifespan (not necessary)
dt[, lif := xmv(lifespan)]
print(dt)
# Anonymously (more useful)
dt[, x2 := x-xmv(ref_x)]
print(dt)
}
