% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavr.R
\name{behavr}
\alias{behavr}
\alias{behavr}
\alias{is.behavr}
\title{todo}
\usage{
behavr(x, metadata)

is.behavr(x)
}
\arguments{
\item{x}{\link{data.table} containing all measurments}

\item{metadata}{\link{data.table} containing the metadata}
}
\description{
todo
}
\details{
Both \code{x} and \code{metadata} should have a \strong{column as a key} with \strong{the same name} (typically named \code{id}).
}
\examples{
set.seed(1)
met <- data.table::data.table(id = 1:5,
                              condition=letters[1:5],
                              sex=c("M","M","M","F", "F"),
                              key="id")
data <- met[  ,
              list(t=1L:100L,
                  x=rnorm(100),
                  y=rnorm(100),
                  eating=runif(100) > .5 ),
              by="id"]

d <- behavr(data,met)
print(d)
summary(d)
}
