% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{bin_all}
\alias{bin_all}
\title{Convenience wrapper around bin_apply_all and merge_behavr_all}
\usage{
bin_all(
  data,
  x = "t",
  y,
  x_bin_length = 30 * 60,
  FUN = mean,
  keep_columns = NULL
)
}
\description{
Bin variable \code{y} along variable \code{x} on bins of length \code{x_bin_length}
using function \code{summary_FUN}.
}
