% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin-apply.R
\name{bin_all}
\alias{bin_all}
\title{Convenience wrapper around bin_apply_all and merge_behavr_all}
\usage{
bin_all(data, y, x = "t", ...)
}
\arguments{
\item{data}{A behavr table}

\item{y}{Columns in data to compute aggregates for. Can be more than 1!}

\item{x}{Name of column to make bins of, normally "t" for time}

\item{...}{Additional arguments to bin_apply_all}
}
\description{
Bin variable \code{y} along variable \code{x} on bins of length \code{x_bin_length}
using function \code{summary_FUN}.
}
