% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rejoin.R
\name{rejoin}
\alias{rejoin}
\alias{unjoin}
\title{Join data and metadata}
\usage{
rejoin(x)

unjoin(b, dt)
}
\arguments{
\item{x}{\link{behavr} object}

\item{b}{Reference behavr dataset}

\item{dt}{data.table resulting from a previously rejoined behavr}
}
\value{
a \link{data.table}
}
\description{
This function joins the data of a \link{behavr} table to its own metadata.
When dealing with large data sets, it is preferable to keep metadata and data separate
until a summary of data is computed.
Indeed, joining many metavariables to
very long time series may result in unnecessary -- and prohibitively -- large memory footprint.
}
\details{
This function is useful if dt has undergone some processing / filtering
in its data.table form but we want to recover again its behavr original format
}
\examples{
set.seed(1)
met <- data.table::data.table(id = 1:5,
                              condition = letters[1:5],
                              sex = c("M", "M", "M", "F", "F"),
                              key = "id")
data <- met[,
             list(t = 1L:100L,
                  x = rnorm(100),
                  y = rnorm(100),
                  eating = runif(100) > .5 ),
             by = "id"]

d <- behavr(data, met)
summary_d <- d[, .(test = mean(x)), by = id]
rejoin(summary_d)
# add some meta variable to the metadata
# to showcase that the metadata columns are also recovered
dt <- behavr::toy_ethoscope_data()
metadata <- behavr::meta(dt)
metadata[, metavar := "C"]
behavr::setmeta(dt, metadata)
dt_rejoined <- behavr::rejoin(dt)
dt_unjoin  <- behavr::split(dt_rejoined)
}
\seealso{
\itemize{
\item \link{behavr} -- to formally create a behavr object
}
}
