% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy-data.R
\name{toy_activity_data}
\alias{toy_activity_data}
\alias{toy_ethoscope_data}
\alias{toy_dam_data}
\title{Generate toy activity and sleep data mimiking Drosophila behaviour in tubes}
\usage{
toy_activity_data(query = NULL, seed = 1, rate_range = 1/c(60, 10),
  duration = days(5), sampling_period = 10, ...)

toy_ethoscope_data(...)

toy_dam_data(...)
}
\arguments{
\item{query}{query (i.e. a dataframe where every row defines an animal).
Typically queries have, at least, the columns \code{experiment_id} and \code{region_id}.
The default value (\code{NULL}), will generate data for a single animal.}

\item{seed}{random seed used (see \link{set.seed})}

\item{rate_range}{a parameter defining the boundaries of rate at which animals wake up.
It will be uniformely distributed between animals, but fixed for each animal.}

\item{duration}{length (in seconds) of the data to generate}

\item{sampling_period}{sampling period (in seconds) of the resulting data}

\item{...}{additional arguments to be passed to \code{simulate_animal_activity}}
}
\value{
a \link{behavr} table with the query columns as metavariables.
In addition to \code{id} and \code{t} columns different methods will output different variables:
\itemize{
\item \code{toy_activity_data} will have \code{asleep} and \code{moving} (1/10s)
\item \code{toy_dam_data} will have \code{activity} (1/60s)
\item \code{toy_ethoscope_data} will have \code{xy_dist_log10x1000}, \code{has_interacted} and \code{x} (2/1s)
}
}
\description{
This function generates random data that emulates some of the features of fruit fly activity and sleep.
This is designed \strong{exclusively to provide material for examples and tests} as it generates "realistic" datasets of arbitrary length.
}
\examples{
# just one animal, no query needed
dt <- toy_ethoscope_data(duration = days(3))

# advanced, using a query
query<- data.frame(experiment_id = "toy_experiment",
                   region_id = 1:10,
                   condition = c("A", "B"))


# Data that could come from loadEthoscopeData:
dt <- toy_ethoscope_data(query, duration = days(1))
print(dt)

# Some DAM-like data
dt <- toy_dam_data(query, seed = 2, duration = days(3))
print(dt)

# data where behaviour is annotated e.g. by a classifier
dt <- toy_activity_data(query, 3)
print(dt)
}
